gemini: partition key generation improved
A dedictated source concept is introduced and producing and consuming
partition keys are both coordinated through this object. The size of
the buffers governing how many partition keys are buffered can be
controlled through the CLI args `partition-key-buffer-size` and
`partition-key-buffer-reuse-size` respectively.

The `concurrency` concept is somewhat changed into meaning that
each `job` is running with this number of goroutines. So for
a `mixed` mode it means that `2*concurrency` number of goroutines
will be running at the same time. You may need to rescale your
command line arg `concurrency` accordingly.
