+++ routing_key.go
@@ -7,16 +7,6 @@ import (
 	"github.com/gocql/gocql"
 )
 
-const (
-	protoDirectionMask = 0x80
-	protoVersionMask   = 0x7F
-	protoVersion1      = 0x01
-	protoVersion2      = 0x02
-	protoVersion3      = 0x03
-	protoVersion4      = 0x04
-	protoVersion5      = 0x05
-)
-
 type RoutingKeyCreator struct {
 	routingKeyBuffer []byte
 }
